Move `Robot` class up one level to "MapObjects" folder
The `Robot` class is an abstract base, much like `MapObject` is. It would be better to place the abstract base classes together, for ease of reference, and all the derived classes, which can actually be instantiated, should be kept together in the "Robots" folder.
