+++ OPHD/MapObjects/Robot.cpp
