+++ OPHD/MapObjects/Robots/Robodigger.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Robot.h"
+#include "../Robot.h"
 
 #include "../../Common.h"
 #include "../../Constants/Strings.h"
