+++ OPHD/MapObjects/Robots/Robominer.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Robot.h"
+#include "../Robot.h"
 #include "../../Constants/Strings.h"
