+++ OPHD/UI/MiniMap.cpp
@@ -3,7 +3,7 @@
 #include "../Cache.h"
 #include "../Map/TileMap.h"
 #include "../Map/MapView.h"
-#include "../MapObjects/Robots/Robot.h"
+#include "../MapObjects/Robot.h"
 #include "../States/Route.h"
 #include "../StructureManager.h"
